+++ vllm_rbln/v1/worker/optimum_worker.py
@@ -32,9 +32,11 @@ from vllm.v1.core.sched.output import GrammarOutput, SchedulerOutput
 from vllm.v1.kv_cache_interface import KVCacheSpec
 from vllm.v1.outputs import AsyncModelRunnerOutput, ModelRunnerOutput
 from vllm.v1.worker.worker_base import WorkerBase
-from vllm_rbln.utils.optimum.configuration import update_vllm_num_blocks, get_rbln_params
+
 import vllm_rbln.rbln_envs as envs
 from vllm_rbln.logger import init_logger
+from vllm_rbln.utils.optimum.cache_blocks import sync_num_blocks
+from vllm_rbln.utils.optimum.rbln_params import get_rbln_params
 from vllm_rbln.v1.worker.optimum_model_runner import RBLNOptimumModelRunner
 
 logger = init_logger(__name__)
@@ -113,17 +115,13 @@ class RBLNOptimumWorker(WorkerBase):
         adapter = self.model_runner.model.kv_block_adapter
         num_gpu_blocks = adapter.get_available_num_blocks()
         # If the model is compiled in the runner,
-        # the number of blocks is not set in the vLLM config.
+        # the number of blocks is not set in the vLLM config yet.
         # Therefore, we need to update it here.
         if not self.model_runner.vllm_config.cache_config.num_gpu_blocks:
             num_blocks, _, _, _, _ = get_rbln_params(
-                self.model_runner.vllm_config,
-                self.model_runner.model.rbln_model_config
+                self.model_runner.vllm_config, self.model_runner.model.rbln_model_config
             )
-            update_vllm_num_blocks(
-                self.model_runner.vllm_config,
-                num_blocks
-            )
+            sync_num_blocks(self.model_runner.vllm_config, num_blocks)
 
         validation_blocks = self.model_runner.vllm_config.cache_config.num_gpu_blocks
         # This will be removed after validation check
